Add pet calculator with resource price management
- Implement pet experience calculator with level-based XP requirements
- Add resource price update functionality with async DB save
- Create Stimulus controller for dynamic price updates
- Add styled success/error messages with auto-hide
- Update database schema to use integer for resource prices
- Add single resource (Croquette) with 500 XP value
- Include translations for EN, ES, FR
- Add refresh icon with tooltip for price updates

+++ db/migrate/20251004221323_create_resources.rb
@@ -4,7 +4,7 @@ class CreateResources < ActiveRecord::Migration[8.0]
       t.string :name
       t.string :resource_type
       t.integer :xp_value
-      t.decimal :average_price
+      t.integer :average_price
       t.text :description
 
       t.timestamps
